USBOTG-GADGET-CONFIG: improvements
- increase entropy on MAC address
Current MAC address on host side is computed from the six most
significant bytes of the STM32MP1 unique serial number. This can
cause collision between two boards with same upper bytes in the
unique serial number, when plugged to the same PC
Plus, today, most of STM32MP1 have the most significant byte equal
to zero, and this makes the MAC address valid. But this will not
last forever, and current approach will fail when the byte will be
equal to one, triggering a "broadcast" type MAC address (check
image https://en.wikipedia.org/wiki/File:MAC-48_Address.svg
and description in https://en.wikipedia.org/wiki/MAC_address
that will be rejected by a Linux host PC
Moreover, also the second bit of the most significant byte can
create problems. If it forces a "locally administered" MAC, this
causes the Linux kernel on the host PC to reject it and replace it
with a random MAC locally generated

-  fix multicast issue
Generate the MAC from the unique serial number passed through a
hash function. This guaranties a better randomization of the
generated MAC by taking in account all the bits of the unique
serial number, while still keeps the property of producing always
the same MAC from the same unique serial number
Filter the bits 0 and 1 of the top first byte of the MAC to force
"unicast" and "globally unique" MAC value.

- set USB device serial number from STM32MP1
To identify the STM32MP1 board connected to the PC, reuse the
unique serial number of the STM32MP1 as serial number of the USB
device.
Strip the trailing null byte to create a valid string.

Change-Id: I66f613bcd739e9b46fbd09506ee90a0046850c6e
Signed-off-by: Antonio Borneo <antonio.borneo@st.com>