LTRACE: Add support for ARM Thumb branches with instruction-set exchange.
This is issue has been detected on OpenSTLinux 2019-10-09, where
ltrace failed to single-step on a "bx lr" instruction when tracing
weston-simple-egl.

The original maintainer of ltrace has been contacted without success,
and the projet hosted on Github is not active anymore.  As a
consequence, this patch should be submitted to the Yocto project.

Change-Id: Ifaa9ca0a025fe2ee05b34468588de0c0e0eea24f