DEMO-AI : improve display and input detection
a) automatic input detection : if both mouse and touchscreen are detected,
use touchscreen instead of mouse.
If mouse is not plugged, touchscreen is default input.

b) automatic display resolution : the display output is read
from the connected device from modetest command, and resolution
is read from /sys/class/drm/card0-<display>.

The mouse returns relative coordinates, which needs to be
converted into absolutes, and after a while, an offset
between cursor given by Wayland and the computed coordinates could
happend, especially when cursor wants to go outside the screen.
That's why mouse cursor is initialized at each draw.

An improvement proposition would to use libinput which is the
layer between kernel and Wayland which manages inputs, but the
application will depend on Wayland.
Here is a piece of code as example if we want to go on that direction :
https://github.com/wayland-project/libinput/blob/master/test/test-touch.c

Signed-off-by: Christophe GUIBOUT <christophe.guibout@st.com>
Change-Id: I0d74261334d6045a1b9b6d74a9714879cd298be5